% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_cases_annual.R
\name{analyze_cases_annual}
\alias{analyze_cases_annual}
\title{Generates year by year epicurve}
\usage{
analyze_cases_annual(
  disease_latest,
  to_path = "analyze_cases_annual - Output.png"
)
}
\arguments{
\item{disease_latest}{latest table- requires proxy_onset_date, adj_mw_...}

\item{to_path}{file path for .png}
}
\description{
Creates epicurve with bars for current year cases, and line for previous year cases. Automatically adjusts displayed years based on current morbidity week, keeps cases in the middle, and shows previous years if needed.
}
