% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_cases_regions.R
\name{analyze_cases_regions}
\alias{analyze_cases_regions}
\title{Generate regional analysis facet graph.}
\usage{
analyze_cases_regions(linelist_latest, to_path)
}
\arguments{
\item{linelist_latest}{dataframe of disease table (typically uses latest table)}

\item{to_path}{string, path of generated .png file}
}
\description{
Generates facet graph for regional case breakdown. Uses current year, and previous 2 years. Must contain clean region, adj_mw_..., proxy_onset_date columns. Exports generated to .png file based on to_path argument.
}
