% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_acronym.R
\name{check_acronym}
\alias{check_acronym}
\title{Check for acronyms in strings}
\usage{
check_acronym(strings_vector, target_acronym)
}
\arguments{
\item{strings_vector}{a string or a vector of strings to look from acronyms from}

\item{target_acronym}{acronym to look for. only include letters here, periods for each letter are not needed. (e.g. "JE", not "J.E.")}
}
\value{
a boolean or a vector of booleans for inputted string or strings
}
\description{
returns a boolean or a vector of booleans that checks for acronyms. note that arguments are all case sensitive, and it is recommended to only use capital letters for all arguments.
}
\examples{
check_acronym("WITH J.E.", "JE")
check_acronym(c("WITH HFMD", "HAS HFMD", "NONE"), "HFMD")
}
