% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_soundex_match.R
\name{check_soundex_match}
\alias{check_soundex_match}
\title{Title}
\usage{
check_soundex_match(strings_vector, target_word)
}
\arguments{
\item{strings_vector}{Vector or column to search soundex from}

\item{target_word}{Target word, will search for actual word or matching soundex code}
}
\value{
Vector of booleans, for search output from strings_vector
}
\description{
Title
}
