% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_soundex_match_df.R
\name{check_soundex_match_df}
\alias{check_soundex_match_df}
\title{Title}
\usage{
check_soundex_match_df(df, column_name, target_word)
}
\arguments{
\item{df}{Dataframe to edit; final dataframe will be returned as output}

\item{column_name}{Column name to search from}

\item{target_word}{word to search for}
}
\value{
original argument df, but with added column called "matching_words"
}
\description{
Title
}
