% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_morbidity_weeks.R
\name{download_morbidity_weeks}
\alias{download_morbidity_weeks}
\title{Downloads morbidity weeks.}
\usage{
download_morbidity_weeks()
}
\value{
dataframe, sourced from morbidity weeks google sheet.
}
\description{
Downloads the morbidity week reference from Google Sheet. Sheet contains year, week number, start date, and end date information for each morbidity week.
}
