% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_populations.R
\name{download_populations}
\alias{download_populations}
\title{Downloads population}
\usage{
download_populations()
}
\value{
dataframe, sourced from population google sheet.
}
\description{
Downloads the population reference from Google Sheet. Sheet contains year, provcity, region, population information for each province/HUC.
}
