% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_edcs.R
\name{extract_edcs}
\alias{extract_edcs}
\title{Extracts from EDCS modified within the week previous to the run_date.}
\usage{
extract_edcs(bq_project, bq_edcs_dataset, bq_disease, run_date = Sys.Date())
}
\arguments{
\item{bq_project}{BigQuery project where the data is found.}

\item{bq_edcs_dataset}{BigQuery dataset where EDCS data is found.}

\item{bq_disease}{Code of disease on BigQuery.}

\item{run_date}{Optional. Date to base the morbidity week info off of. If none is given, then the current date will be used.}
}
\value{
A dataframe that left joins the EDCS patient and lab modules. All the information from the patient module is maintained, and only the unique lab module columns are joined.
}
\description{
Extracts from EDCS modified within the week previous to the run_date. Extracts both EDCS and EDCS lab profiles that were last modified in the week before the given run_date.
}
