% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_morbidity_week_shard.R
\name{get_date_shard}
\alias{get_date_shard}
\title{Gets table shard for given date, based on morbidity week}
\usage{
get_date_shard(from_date = Sys.Date(), morbidity_weeks = NULL)
}
\arguments{
\item{from_date}{date to generate shard from, if null, current date is used}

\item{morbidity_weeks}{optional, morbidity week reference from tbandothers::download_morbidity_weeks()}
}
\value{
string of table date shard
}
\description{
Gets table shard for given date, based on morbidity week
}
