% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_morbidity_week_end.R
\name{get_morbidity_week_end}
\alias{get_morbidity_week_end}
\title{Gets the last date of a given morbidity week.}
\usage{
get_morbidity_week_end(week_number, from_year = NULL)
}
\arguments{
\item{week_number}{Desired week number.}

\item{from_year}{Optional. Which year's morbidity week calendar to reference. If not provided, then the current year is used.}
}
\value{
Date
}
\description{
Gets the last date of a given morbidity week.
}
