% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_population.R
\name{get_population}
\alias{get_population}
\title{Gets the FHSIS population estimate given the year, region, and province/HUC.}
\usage{
get_population(to_year, to_region, to_provcity)
}
\arguments{
\item{to_year}{year of population estimates}

\item{to_region}{region of population estimates}

\item{to_provcity}{province/HUC of population estimates}
}
\value{
integer of popoulation
}
\description{
Gets the FHSIS population estimate given the year, region, and province/HUC.
}
