% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\docType{data}
\name{morbidity_weeks}
\alias{morbidity_weeks}
\title{read morbidity week Google Sheet - MUST INSTALL googlesheets4 package first}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 418 rows and 4 columns.
}
\usage{
morbidity_weeks
}
\description{
read morbidity week Google Sheet - MUST INSTALL googlesheets4 package first
}
\keyword{datasets}
