% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_morbidity_weeks.R
\name{read_morbidity_weeks}
\alias{read_morbidity_weeks}
\title{Reads TB and Others morbidity week Google Sheet.}
\usage{
read_morbidity_weeks()
}
\value{
A data frame of all stored morbidity weeks.
}
\description{
Reads TB and Others morbidity week Google Sheet. Only authorized when first ran during session. No confidential information is stored in the Google Sheet.
}
