% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_morbidity_week.R
\name{adjust_morbidity_week}
\alias{adjust_morbidity_week}
\title{Gets the morbidity weeks of a date or dates adjusted to a provided year.}
\usage{
adjust_morbidity_week(dates, to_year)
}
\arguments{
\item{dates}{Either a date or a vector of dates.}

\item{to_year}{The year that the years are based on.}
}
\value{
A vector of morbidity weeks adjusted to provided year.
}
\description{
Gets the morbidity weeks of a date or dates adjusted from a provided year's morbidity week calendar.
}
