% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_morbidity_week.R
\name{get_morbidity_week}
\alias{get_morbidity_week}
\title{Gets morbidity week of a given date.}
\usage{
get_morbidity_week(date)
}
\arguments{
\item{date}{the date which we get the morbidity week for}
}
\value{
Integer of morbidity week
}
\description{
Gets morbidity week of a given date.
}
