% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_morbidity_week.R
\name{get_morbidity_week}
\alias{get_morbidity_week}
\title{Gets morbidity week of either a single date, or a vector of dates.}
\usage{
get_morbidity_week(date)
}
\arguments{
\item{date}{A single date, or vector of dates which we get the morbidity week/s for.}
}
\value{
If single date, then integer of morbidity week of said date. If vector of dates, then vector of morbidity weeks of said dates.
}
\description{
Gets morbidity week of either a single date, or a vector of dates.
}
